Initial version
Now it supports a dump from the hosted version without changes. 
This version has not been testes with wikis or forums.
Also there were several columns that came from the dump that I could not match to anything so they were named tmp_N and are ignored later by the data migration.